Refactor pop_back() logic to remove redundancy
This commit first removes a redundant test in lib.rs that tested
push_front multiple times but did not do anything useful with that
information.  This was a temporary test to see that multiple pushes
could be done without the program panicking.

This commit mainly ensures that the links between nodes are correct
after popping.  Unlike pushing, it turns out the popping logic does
not require a special case when going from 2 nodes to 1.  By manually
testing this, I noticed that the pop_back logic redundantly sets
tail->prev->next to a weak link to the head twice.  Removing this
makes the code cleaner and more efficient.

Furthermore, there is now a test that tests many pushes and pops from
either end of the list, ensuring the links are traversed and reset
accurately.