Add peek_front() and peek_back() methods
This commit allows the user to peek the item at the front or back of
the list.  If the list is not empty, they can unwrap and dereference
the returned Ref<T> to access the value.

I add a few more unit tests to lib.rs to test that peek() works.
Thankfully, this was a simple feature to implement!