Complete insert_at() and implement remove_at()
This commit modifies insert_at() to traverse the list based on the
insertion point.  The same functionality is given to the remove_at()
function, which removes an element from a specific point in the list.
Whereas insert_at() adds an element to the list, remove_at() breaks
the links to the object in the list and returns it to the user.

Further testing is needed to ensure both methods work properly.